Split account linking commands
